Updated the Added Sitemap with dynamic update for SEO
