+++ app/sitemap.ts
@@ -1,7 +1,29 @@
 // app/(site)/sitemap.ts
 import { MetadataRoute } from "next";
+import { createClient } from "@/utils/supabase/server";
 
-export default function sitemap(): MetadataRoute.Sitemap {
+export default async function sitemap(): Promise<MetadataRoute.Sitemap> {
+
+  const supabaseClient = createClient();
+  const { data: listings, count } = await supabaseClient
+    .from('listings')
+    .select('slug, updated_at', { count: 'exact' })
+    .eq('moderation_status', 'approved');
+
+  if (!listings) {
+    return [];
+  }
+
+  // Sort listings by updated_at in descending order
+  listings.sort((a, b) => new Date(b.updated_at).getTime() - new Date(a.updated_at).getTime());
+
+  const listingSitemapEntries: MetadataRoute.Sitemap = listings.map((listing) => ({
+    url: `${process.env.NEXT_PUBLIC_BASE_URL}/directory/${listing.slug}`,
+    lastModified: listing.updated_at,
+    priority: 0.8,
+  }));
+
+
   return [
     {
       url: `${process.env.NEXT_PUBLIC_BASE_URL}`,
@@ -14,12 +36,6 @@ export default function sitemap(): MetadataRoute.Sitemap {
       lastModified: new Date(),
       changeFrequency: "daily",
       priority: 1,
-    },
-    {
-      url: `${process.env.NEXT_PUBLIC_BASE_URL}/directory/sitemap.xml`,
-      lastModified: new Date(),
-      changeFrequency: "daily",
-      priority: 0.8,
     },
     {
       url: `${process.env.NEXT_PUBLIC_BASE_URL}/earn`,
@@ -51,6 +67,7 @@ export default function sitemap(): MetadataRoute.Sitemap {
       changeFrequency: "yearly",
       priority: 1,
     },
+    ...listingSitemapEntries,
     // Add more static URLs here
   ];
 }
